requirements.txt
asn1crypto==1.0.1
backcall==0.1.0
beautifulsoup4==4.8.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
credentials==1.1
cryptography==2.7
decorator==4.4.0
facenet-pytorch==0.3.1
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.13.2
lxml==4.4.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.2
olefile==0.46
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
praw==6.4.0
prawcore==1.0.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
PySocks==1.7.1
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
soupsieve==1.9.3
torch==1.3.0
torchvision==0.4.1a0+d94043a
tqdm==4.36.1
traitlets==4.3.3
update-checker==0.16
urllib3==1.24.2
wcwidth==0.1.7
websocket-client==0.56.0
yml==0.0.1
